Rewrite .obj file loading + bug fixes
There was a weird bug with fscanf() in the previous version,
where the function would consistently fail at certain
lines of the file suzanne.obj. This bug was not my fault
as the same code worked when compiled with Visual Studio.

Also, the way the flags work has been changed. The previous
method that utilized enums was just plain wrong as it
ended up setting the enum variable to values other than
the ones defined in the enum definition..

Experimental back-face culling has also been implemented.
It's not that useful though, as the transformation of the
normals is an expensive operation.. The rendering ends up
slower with the culling enabled than without it. Not ideal.